Fixes the bug where the test_complete signal was emitting the wrong type and causing the HVPSReport to fail.
Fixes the bug where the QLineEdits would not initialize as disabled.

Fixs the bug where pressing enter in the neg 1k measurement QLineEdit would not reenable the other HV test buttons.

Fixes the bug where the focus was not going to the next HV button when enter was pressed in the L1 QLineEdits.